backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==4.2.13
django-allauth==0.63.3
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
mysqlclient==2.2.4
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.0
typing-extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
